+++ TABP.Application/Queries/Hotels/SearchAndFilterHotelsQuery.cs
@@ -16,6 +16,6 @@ public class SearchAndFilterHotelsQuery : IRequest<List<HotelSearchResponseDto>>
     public string? HotelName { get; set; }
     public string? City { get; set; }
     public decimal? MaxPrice { get; init; }
-    public int? MinRating { get; init; }
+    public float? MinRating { get; init; }
     public List<Guid>? Amenities { get; init; }
 }
